backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-groq==0.3.5
langchain-huggingface==0.3.0

faiss-cpu==1.11.0
sentence-transformers==5.0.0
transformers==4.53.1

pypdf==5.7.0

redis==5.0.3
python-dotenv==1.1.1

requests==2.32.4
groq==0.29.0